pyproject.toml
[tool.poetry]
name = "led-cube-data"
version = "0.1.0"
description = "File specification used in my LED cube projects."
authors = ["crash8229 <mu304007@gmail.com>"]
license = "GPL-3.0-only"
readme = "README.md"
packages = [{include = "led_cube_data"}]
repository = "https://github.com/crash8229/led-cube-data"
homepage = "https://github.com/crash8229/led-cube-data"
classifiers = ["Development Status :: 4 - Beta"]

[tool.poetry.dependencies]
python = "^3.10"
kaitaistruct = "^0.10"
PyYAML = "^6.0"
construct = "^2.10.68"

[tool.poetry.group.dev.dependencies]
construct-typing = "^0.5.3"
pytest = "^7.0"
black = "^22.3.0"
mypy = "^0.982"
types-PyYAML = "^6.0.7"
types-pkg-resources = "^0.1.3"
pytest_check = "^1.0.5"
coverage = "^6.4.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
